Refactor shared models to src/orchestrator/models.py and fix hacky import.
Consolidated TestResult, LanguagePerformance, TestAnalysis, OverallRankings, and PerformanceSummary into a single models.py file to resolve circular dependencies and replace hacky inline imports. Updated core, results, runners, and reports to use the shared models and added defensive psutil imports to metrics and runners.
