+++ src/orchestrator/metrics.py
@@ -6,7 +6,10 @@ Monitors system resources and execution metrics during benchmark tests.
 import os
 import sys
 import time
-import psutil
+try:
+    import psutil
+except ImportError:
+    psutil = None
 import threading
 from typing import Dict, List, Optional, Any, Callable
 from dataclasses import dataclass, field
@@ -65,6 +68,8 @@ class MetricsCollector:
     
     def __init__(self, system_config: SystemConfig):
         self.config = system_config
+        if psutil is None:
+            self.config.monitor_resources = False
         self.monitoring_active = False
         self.monitoring_thread: Optional[threading.Thread] = None
         self.current_metrics: Optional[ExecutionMetrics] = None
@@ -90,7 +95,7 @@ class MetricsCollector:
         self.process_metrics_history.clear()
         
         # Set target process if provided
-        if process_id:
+        if process_id and psutil:
             try:
                 self.target_process = psutil.Process(process_id)
             except psutil.NoSuchProcess:
@@ -147,6 +152,8 @@ class MetricsCollector:
     
     def _collect_system_metrics(self) -> Optional[SystemMetrics]:
         """Collect current system metrics."""
+        if not psutil:
+            return None
         try:
             # CPU and memory
             cpu_percent = psutil.cpu_percent(interval=None)
@@ -257,7 +264,7 @@ class MetricsCollector:
     
     def get_current_system_info(self) -> Dict[str, Any]:
         """Get current system information."""
-        if not self.config.collect_system_info:
+        if not self.config.collect_system_info or not psutil:
             return {}
         
         try:
@@ -358,10 +365,13 @@ class SimpleMetricsCollector:
     def start(self) -> None:
         """Start timing."""
         self.start_time = time.perf_counter()
-        try:
-            process = psutil.Process()
-            self.peak_memory = process.memory_info().rss
-        except Exception:
+        if psutil:
+            try:
+                process = psutil.Process()
+                self.peak_memory = process.memory_info().rss
+            except Exception:
+                self.peak_memory = 0
+        else:
             self.peak_memory = 0
     
     def stop(self) -> Dict[str, float]:
@@ -372,12 +382,13 @@ class SimpleMetricsCollector:
         if self.start_time:
             execution_time = self.end_time - self.start_time
         
-        try:
-            process = psutil.Process()
-            current_memory = process.memory_info().rss
-            self.peak_memory = max(self.peak_memory, current_memory)
-        except Exception:
-            pass
+        if psutil:
+            try:
+                process = psutil.Process()
+                current_memory = process.memory_info().rss
+                self.peak_memory = max(self.peak_memory, current_memory)
+            except Exception:
+                pass
         
         return {
             'execution_time': execution_time,
